requirements.txt
azure-functions==1.21.3
python-dotenv==1.0.1

# gRPC stack pinned to versions compatible with Azure Functions Core Tools
grpcio==1.54.0
grpcio-tools==1.54.0
protobuf==4.23.4

# Auth and crypto
PyJWT[crypto]==2.9.0
cryptography==43.0.1

# HTTP and serialization
requests==2.32.3
avro==1.12.0
fastavro==1.9.7

# Snowflake connector
snowflake-connector-python==3.12.2

